backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiofiles==23.2.1
redis==5.0.4
httpx==0.27.0
websockets==12.0
pywebpush==2.0.0
pillow==10.3.0
python-dotenv==1.0.1
emails==0.6
